+++ tests/filters.py
@@ -38,6 +38,34 @@ DEFAULT_FILTERS = {
     },
 }
 
+RENAMED_EXPANDABLE_MODEL_FIELDS = {
+    str: ALWAYS,
+    int: ALWAYS,
+    dict: ALWAYS,
+    models.TestModel: {
+        "pk": ALWAYS,
+        "int_field": ALWAYS,
+        "renamed_expandable_dict": expandable(inst_field_name="expandable_dict"),
+        "renamed_expandable_string": expandable(inst_field_name="expandable_string"),
+    },
+    models.ParentModel: {
+        "nonstandard_id": ALWAYS,
+        "name": ALWAYS,
+        "favorite": expandable(model_class=models.ChildModel, display_key="name"),
+        "children": expandable(model_class=models.ChildModel, display_key="name"),
+    },
+    models.ChildModel: {
+        "pk": ALWAYS,
+        "renamed_test": expandable(
+            model_class=models.TestModel, inst_field_name="test"
+        ),
+        "name": ALWAYS,
+        "renamed_parent": expandable(
+            model_class=models.ParentModel, inst_field_name="parent"
+        ),
+    },
+}
+
 DEFAULT_FILTERS_NO_EXPANDABLE = {
     str: ALWAYS,
     int: ALWAYS,