+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+# [Unrelased]
+### Changed
+- [PR 39](https://github.com/salesforce/django-declarative-apis/pull/39) Add long_description to setup 
+
 # [0.20.0] - 2020-03-09
 ### Added
 - [PR 21](https://github.com/salesforce/django-declarative-apis/pull/21) Add Makefile targets for Black formatter